fix(imgproc): address Qodo review comments for SIMD compare_hist
- Moves SIMD helper from core to imgproc/simd.rs.
- Fixes numerical overflow by casting intermediate f32 to f64 for accumulations.
- Fixes epsilon classification of small f32 bins by using f64::EPSILON.
- Restores exact f64::min() behavior for Intersection method to handle NaNs correctly.
- Adds tests.rs coverage for compare_hist to satisfy Qodo rules.
