+++ src/core/simd.rs
@@ -189,14 +189,6 @@ pub trait SimdElement: Copy + Send + Sync + 'static {
         false
     }
 
-    // -- Histogram comparisons --
-
-    /// Compute compare_hist directly using the method variant mapped to a `u8`.
-    /// 0=Correl, 1=ChiSqr, 2=ChiSqrAlt, 3=Intersection, 4=Bhattacharyya, 5=KullbackLeibler
-    fn simd_compare_hist(_h1: &[Self], _h2: &[Self], _method: u8) -> Option<f64> {
-        None
-    }
-
     // -- Pyramid kernels (5-tap Gaussian) --
 
     /// Horizontal pass: dst[i] = sum(src[i - 2*stride .. i + 2*stride] * [1, 4, 6, 4, 1])
@@ -502,108 +494,6 @@ mod simd_impls {
                 _ => return false,
             }
             true
-        }
-
-        fn simd_compare_hist(h1: &[Self], h2: &[Self], method: u8) -> Option<f64> {
-            match method {
-                0 => {
-                    // Correl
-                    let mut s1 = 0.0f32;
-                    let mut s2 = 0.0f32;
-                    let mut s11 = 0.0f32;
-                    let mut s12 = 0.0f32;
-                    let mut s22 = 0.0f32;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        s1 += a;
-                        s2 += b;
-                        s11 += a * a;
-                        s22 += b * b;
-                        s12 += a * b;
-                    }
-                    let n = h1.len() as f64;
-                    let scale = 1.0 / n;
-                    let s1_f = s1 as f64;
-                    let s2_f = s2 as f64;
-                    let s11_f = s11 as f64;
-                    let s12_f = s12 as f64;
-                    let s22_f = s22 as f64;
-
-                    let num = s12_f - s1_f * s2_f * scale;
-                    let denom2 = (s11_f - s1_f * s1_f * scale) * (s22_f - s2_f * s2_f * scale);
-                    Some(if denom2.abs() > f64::EPSILON {
-                        num / denom2.sqrt()
-                    } else {
-                        1.0
-                    })
-                }
-                1 => {
-                    // ChiSqr
-                    let mut result = 0.0f32;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        let diff = a - b;
-                        let a_adj = if a.abs() <= f32::EPSILON { 1.0 } else { a };
-                        let val = diff * diff / a_adj;
-                        let add_val = if a.abs() > f32::EPSILON { val } else { 0.0 };
-                        result += add_val;
-                    }
-                    Some(result as f64)
-                }
-                2 => {
-                    // ChiSqrAlt
-                    let mut result = 0.0f32;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        let sum = a + b;
-                        let diff = a - b;
-                        let sum_adj = if sum.abs() <= f32::EPSILON { 1.0 } else { sum };
-                        let val = diff * diff / sum_adj;
-                        let add_val = if sum.abs() > f32::EPSILON { val } else { 0.0 };
-                        result += add_val;
-                    }
-                    Some((result as f64) * 2.0)
-                }
-                3 => {
-                    // Intersection
-                    let mut result = 0.0f32;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        result += if a < b { a } else { b };
-                    }
-                    Some(result as f64)
-                }
-                4 => {
-                    // Bhattacharyya
-                    let mut s1 = 0.0f32;
-                    let mut s2 = 0.0f32;
-                    let mut bc = 0.0f32;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        s1 += a;
-                        s2 += b;
-                        bc += (a * b).sqrt();
-                    }
-                    let s1_f = s1 as f64;
-                    let s2_f = s2 as f64;
-                    let bc_f = bc as f64;
-                    let norm = s1_f * s2_f;
-                    let norm_factor = if norm.abs() > f64::EPSILON {
-                        1.0 / norm.sqrt()
-                    } else {
-                        1.0
-                    };
-                    Some(((1.0 - bc_f * norm_factor).max(0.0)).sqrt())
-                }
-                5 => {
-                    // KullbackLeibler
-                    let mut result = 0.0f32;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        let q_adj = if b.abs() <= f32::EPSILON { 1e-10 } else { b };
-                        let p_adj = if a.abs() <= f32::EPSILON { 1.0 } else { a };
-                        let val = a * (p_adj / q_adj).ln();
-                        let add_val = if a.abs() > f32::EPSILON { val } else { 0.0 };
-                        result += add_val;
-                    }
-                    Some(result as f64)
-                }
-                _ => None,
-            }
         }
 
         fn simd_gaussian_5tap_h(dst: &mut [f64], src: &[Self], stride: usize) -> bool {
@@ -1011,100 +901,6 @@ mod simd_impls {
                 _ => return false,
             }
             true
-        }
-
-        fn simd_compare_hist(h1: &[Self], h2: &[Self], method: u8) -> Option<f64> {
-            match method {
-                0 => {
-                    // Correl
-                    let mut s1 = 0.0f64;
-                    let mut s2 = 0.0f64;
-                    let mut s11 = 0.0f64;
-                    let mut s12 = 0.0f64;
-                    let mut s22 = 0.0f64;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        s1 += a;
-                        s2 += b;
-                        s11 += a * a;
-                        s22 += b * b;
-                        s12 += a * b;
-                    }
-                    let n = h1.len() as f64;
-                    let scale = 1.0 / n;
-
-                    let num = s12 - s1 * s2 * scale;
-                    let denom2 = (s11 - s1 * s1 * scale) * (s22 - s2 * s2 * scale);
-                    Some(if denom2.abs() > f64::EPSILON {
-                        num / denom2.sqrt()
-                    } else {
-                        1.0
-                    })
-                }
-                1 => {
-                    // ChiSqr
-                    let mut result = 0.0f64;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        let diff = a - b;
-                        let a_adj = if a.abs() <= f64::EPSILON { 1.0 } else { a };
-                        let val = diff * diff / a_adj;
-                        let add_val = if a.abs() > f64::EPSILON { val } else { 0.0 };
-                        result += add_val;
-                    }
-                    Some(result)
-                }
-                2 => {
-                    // ChiSqrAlt
-                    let mut result = 0.0f64;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        let sum = a + b;
-                        let diff = a - b;
-                        let sum_adj = if sum.abs() <= f64::EPSILON { 1.0 } else { sum };
-                        let val = diff * diff / sum_adj;
-                        let add_val = if sum.abs() > f64::EPSILON { val } else { 0.0 };
-                        result += add_val;
-                    }
-                    Some(result * 2.0)
-                }
-                3 => {
-                    // Intersection
-                    let mut result = 0.0f64;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        result += if a < b { a } else { b };
-                    }
-                    Some(result)
-                }
-                4 => {
-                    // Bhattacharyya
-                    let mut s1 = 0.0f64;
-                    let mut s2 = 0.0f64;
-                    let mut bc = 0.0f64;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        s1 += a;
-                        s2 += b;
-                        bc += (a * b).sqrt();
-                    }
-                    let norm = s1 * s2;
-                    let norm_factor = if norm.abs() > f64::EPSILON {
-                        1.0 / norm.sqrt()
-                    } else {
-                        1.0
-                    };
-                    Some(((1.0 - bc * norm_factor).max(0.0)).sqrt())
-                }
-                5 => {
-                    // KullbackLeibler
-                    let mut result = 0.0f64;
-                    for (a, b) in h1.iter().copied().zip(h2.iter().copied()) {
-                        let q_adj = if b.abs() <= f64::EPSILON { 1e-10 } else { b };
-                        let p_adj = if a.abs() <= f64::EPSILON { 1.0 } else { a };
-                        let val = a * (p_adj / q_adj).ln();
-                        let add_val = if a.abs() > f64::EPSILON { val } else { 0.0 };
-                        result += add_val;
-                    }
-                    Some(result)
-                }
-                _ => None,
-            }
         }
 
         fn simd_gaussian_5tap_h(dst: &mut [f64], src: &[Self], stride: usize) -> bool {
