+++ src/imgproc/histogram.rs
@@ -575,11 +575,10 @@ pub fn compare_hist(h1: &Matrix<f32>, h2: &Matrix<f32>, method: HistCompMethods)
 
     #[cfg(feature = "simd")]
     {
-        use crate::core::simd::SimdElement;
-        if f32::has_simd() {
-            if let Some(res) = f32::simd_compare_hist(&h1.data, &h2.data, method as u8) {
-                return Ok(res);
-            }
+        if let Some(res) =
+            crate::imgproc::simd::simd_compare_hist_f32(&h1.data, &h2.data, method as u8)
+        {
+            return Ok(res);
         }
     }
